+++ tests/testthat/test-merPredict.R
@@ -36,13 +36,15 @@ test_that("Prediction intervals work for simple GLMM example", {
   skip_on_travis()
   skip_on_cran()
   set.seed(101)
-  d <- expand.grid(fac1=LETTERS[1:5], grp=factor(1:10),
-                   obs=1:50)
-  d$y <- simulate(~fac1+(1|grp),family = binomial,
-                  newdata=d,
-                  newparams=list(beta=c(2,-1,3,-2,1.2), theta=c(.33)),
+  d <- expand.grid(fac1 = LETTERS[1:5],
+                   grp = letters[11:20],
+                   obs = 1:50)
+  d$y <- simulate(~fac1 + (1 | grp), family = binomial,
+                  newdata = d,
+                  newparams = list( beta = c(2,-1,3,-2,1.2),
+                                    theta = c(.33)),
                   seed =634)[[1]]
-  subD <- d[sample(row.names(d), 1200),]
+  subD <- d[sample(row.names(d), 1200), ]
 
   g1 <- glmer(y~fac1+(1|grp), data=subD, family = 'binomial')
   d$fitted <- predict(g1, d)